+++ CHANGES.rst
@@ -5,6 +5,7 @@ Version 0.9.2 (UNRELEASED)
 --------------------------
 
 - Adds option to delete all the runs of a workflow.
+- Changes the launch on REANA page to display the optional launch parameters in a table.
 - Changes Docker image Node version from 16 to 18.
 - Fixes container image building on the arm64 architecture.